test(file-tree): harden ESM import coverage
